split shared functionalities into separate file
