+++ flight-controller/controller-code/utilities/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(hafx-spectrum hafx_spectrum.cc)
+add_executable(hafx-spectrum hafx_spectrum.cc common.hh)
 
 target_link_libraries(
     hafx-spectrum
@@ -8,6 +8,21 @@ target_link_libraries(
 
 target_include_directories(
     hafx-spectrum
+    PUBLIC
+        .
+)
+
+
+add_executable(hafx-trace hafx_scope_trace.cc common.hh)
+
+target_link_libraries(
+    hafx-trace
+    PUBLIC
+        hafx-ctrl
+)
+
+target_include_directories(
+    hafx-trace
     PUBLIC
         .
 )
